# swagger.yaml
openapi: 3.0.0
info:
  title: Senior High Grading System API
  description: API for managing senior high school grading system including accounts, students, subjects, enrollments, quizzes, and grades
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.production.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        sex:
          type: string
          enum: [M, F]
        email:
          type: string
          format: email
        homeroom_id:
          type: integer
        address:
          type: string
        school_id:
          type: string
        created:
          type: string
          format: date-time
    Subject:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            [
              Core,
              Applied,
              Specialized_GAS,
              Specialized_ABM,
              Specialized_HUMMS,
              Specialized_STEM,
            ]
    Quiz:
      type: object
      properties:
        id:
          type: integer
        teacher_subject_id:
          type: integer
        type:
          type: string
          enum: [Written Work, Performance Tasks, Quarterly Assessment]
        quarter:
          type: string
          enum: [First Quarter, Second Quarter]
        description:
          type: string
        hps:
          type: number
          description: Highest Possible Score
...
paths:
...
  /students:
    get:
      tags:
        - Students
      summary: Get all students
      description: Retrieve all students (Registrar only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"

    post:
      tags:
        - Students
      summary: Create new student
      description: Register a new student (Registrar only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstName
                - lastName
                - sex
                - email
                - homeroom_id
                - school_id
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                sex:
                  type: string
                  enum: [M, F]
                email:
                  type: string
                  format: email
                homeroom_id:
                  type: integer
                address:
                  type: string
                school_id:
                  type: string
                  maxLength: 20
      responses:
        "200":
          description: Student created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      tags:
        - Students
      summary: Get student information
      description: Retrieve detailed student information
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/egrades/{id}:
    get:
      tags:
        - Students
      summary: Get student subjects and grades
      description: Retrieve student's enrolled subjects and their grades
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student subjects and grades
          content:
            application/json:
              schema:
                type: object
  /subjects:
    get:
      tags:
        - Subjects
      summary: Get all subjects
      description: Retrieve all subjects (Principal only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"

    post:
      tags:
        - Subjects
      summary: Create new subject
      description: Create a new subject (SuperAdmin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - name
                - type
              properties:
                code:
                  type: string
                name:
                  type: string
                type:
                  type: string
                  enum:
                    [
                      Core,
                      Applied,
                      Specialized_GAS,
                      Specialized_ABM,
                      Specialized_HUMMS,
                      Specialized_STEM,
                    ]
      responses:
        "200":
          description: Subject created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
...
  /quizzes:
    post:
      tags:
        - Quizzes
      summary: Add new quiz
      description: Create a new quiz or assessment (Teacher only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - teacher_subject_id
                - type
                - quarter
                - hps
              properties:
                teacher_subject_id:
                  type: integer
                type:
                  type: string
                  enum: [Written Work, Performance Tasks, Quarterly Assessment]
                quarter:
                  type: string
                  enum: [First Quarter, Second Quarter]
                description:
                  type: string
                  maxLength: 255
                hps:
                  type: number
                  minimum: 1
      responses:
        "200":
          description: Quiz created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
  /quizzes/{id}:
    get:
      tags:
        - Quizzes
      summary: Get quizzes
      description: Get quizzes for a teacher subject with optional filters
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Teacher subject ID
        - name: quarter
          in: query
          schema:
            type: string
            enum: [First Quarter, Second Quarter]
        - name: type
          in: query
          schema:
            type: string
            enum: [Written Work, Performance Tasks, Quarterly Assessment]
      responses:
        "200":
          description: List of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"

    put:
      tags:
        - Quizzes
      summary: Update quiz
      description: Update quiz details (Teacher only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - hps
              properties:
                hps:
                  type: number
                  minimum: 1
                description:
                  type: string
                  maxLength: 255
      responses:
        "200":
          description: Quiz updated successfully

    delete:
      tags:
        - Quizzes
      summary: Delete quiz
      description: Delete a quiz (Teacher only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
  /quizzes/quarterly-grade-sheet/{id}:
    get:
      tags:
        - Quizzes
      summary: Get quarterly grade sheet
      description: Get grade sheet for a specific quarter
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Teacher subject ID
        - name: quarter
          in: query
          required: true
          schema:
            type: string
            enum: [First Quarter, Second Quarter]
      responses:
        "200":
          description: Quarterly grade sheet
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /quizzes/semestral-final-grade/{id}:
    get:
      tags:
        - Quizzes
      summary: Get semestral final grade
      description: Get final grades for the semester
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Teacher subject ID
      responses:
        "200":
          description: Semestral final grades
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
  /subject-quarter-locks:
    get:
      tags:
        - Subject Quarter Locks
      summary: Get subject quarter locks
      description: Retrieve subject quarter lock status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of subject quarter locks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...